1:1 translate the WLine smooth approximation chain (GeomInt_WLApprox/ApproxInt) with the OCCT TKMath solvers
- geomalgo/approx_int.rs: 1:1 of ApproxInt_KnotTools.cxx (EvalCurv,
  BuildCurvature, ComputeKnotInds, InsKnotBefI, FilterKnots, BuildKnots,
  DefineParType), ApproxInt_Approx.gxx (prepareDS/fillData/buildKnots/
  buildCurve/SetParameters), Approx_ComputeLine.gxx (Perform/Compute/
  ComputeCurve/CheckMultiCurve/TangencyVector/SearchLambda) and
  AppParCurves_LeastSquare.gxx (SearchIndex/MakeTAA/DACTCL/Householder)
- rcad-kernel math: VecD/MatD/IntVec (math_Vector/Matrix/IntegerVector),
  lin::householder_ls (math_Householder.cxx L85-179), lin::dactcl_*
  (math_Recipes.cxx L750-901), opt::bracket_minimum (math_BracketMinimum
  L62-211), opt::bfgs_minimize_occt (math_BFGS L110-454)
- face_make_curve Walking branch: 1:1 of IntTools_FaceFace.cxx L1175-1391
  with ApproxParameters (L2736-2783) and the MultiBSpCurve extraction
  (L1642-1733); knots mapped to the part range [fprm, lprm]
- surface.rs: Green boundary integral per knot span (C0 junction exactness)
- bopfuse: ze3-ze8/zf1 pass (SA 38.05805 vs 38.0586); bopcommon/bopcut/
  boptuc failure counts drop with no regression

+++ libs/rcad-kernel/src/topo/topods.rs
@@ -3120,14 +3120,14 @@ mod tests {
         let v1 = brep.add_tvertex(DVec3::X);
         let e = bld.add_edge(&mut brep, None, v0.clone(), v1.clone(), [0.0, 1.0]);
         bld.set_vertex_param(&mut brep, e.clone(), v0.clone(), 0.5);
-        assert_eq!(brep.edge(e.clone()).vertex_params.get(&v0.index), Some(&0.5));
+        assert_eq!(brep.edge(e.clone()).vertex_params.get(&(v0.index as u64)), Some(&0.5));
 
         // Transfer from v0 on e to v_new on a new edge
         let v_new = brep.add_tvertex(DVec3::new(2.0, 0.0, 0.0));
         let e_new = bld.add_edge(&mut brep, None, v_new.clone(), v1.clone(), [0.0, 1.0]);
-        assert!(brep.edge(e_new.clone()).vertex_params.get(&v_new.index).is_none());
+        assert!(brep.edge(e_new.clone()).vertex_params.get(&(v_new.index as u64)).is_none());
         bld.transfert_vertex_param(&mut brep, e.clone(), v0.clone(), e_new.clone(), v_new.clone());
-        assert_eq!(brep.edge(e_new.clone()).vertex_params.get(&v_new.index), Some(&0.5));
+        assert_eq!(brep.edge(e_new.clone()).vertex_params.get(&(v_new.index as u64)), Some(&0.5));
     }
 
     #[test]
